requirements.txt
urllib3
requests
pymongo[srv]
beautifulsoup4
python-dotenv
pydantic<2.0
discord.py